+++ metadata.rb
@@ -4,7 +4,8 @@ maintainer_email 'scarman@coldlight.com'
 license          'Apache 2.0'
 description      'Installs/Configures Tomcat'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '0.1.2'
+
+version          '0.2.0'
 
 supports "redhat"
 supports "centos"